Demo of patch support
Related to #1285. Example only, not planning on merging to master.

Unfortunately this makes the JRE warn with "WARNING: Illegal reflective access by org.jsoup.helper.HttpConnection$Response", so not suitable for use.

Best thought  would be to see the missing PATCH support on the underlying HttpURLConnection and swap to a POST with X-HTTP-Method-Override=PATCH. That will help some but not all cases.